backend/requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.15,<3.16
Markdown>=3.6,<4
django-cors-headers>=4.4.0,<4.5
djangorestframework-simplejwt>=5.3.1,<5.4
django-filter>=24.2,<25
gunicorn>=22.0.0,<22.1
python-dotenv>=1.0.1,<1.1
psycopg2-binary>=2.9,<2.10
django-import-export>=4.1,<4.2
openpyxl>=3.1,<3.2